+++ lib/grpc/friends.dart
@@ -1,4 +1,5 @@
 import 'package:fixnum/fixnum.dart' as $fixnum;
+import 'package:grpc/grpc.dart';
 import 'package:grpc/grpc_or_grpcweb.dart';
 import 'package:poll_and_play/config.dart';
 import 'package:poll_and_play/grpc/authenticator.dart';
@@ -8,9 +9,7 @@ import 'package:poll_play_proto_gen/public.dart';
 class FriendsClient {
   late FriendsServiceClient _client;
 
-  FriendsClient(List<String> address) {
-    final channel = GrpcOrGrpcWebClientChannel.toSingleEndpoint(
-        host: address[0], port: int.parse(address[1]), transportSecure: GlobalConfig().secureTransport);
+  FriendsClient(ClientChannel channel) {
     _client = FriendsServiceClient(channel);
   }