+++ lib/grpc/games.dart
@@ -1,3 +1,4 @@
+import 'package:grpc/grpc.dart';
 import 'package:grpc/grpc_or_grpcweb.dart';
 import 'package:poll_and_play/config.dart';
 import 'package:poll_and_play/grpc/authenticator.dart';
@@ -7,9 +8,7 @@ import 'package:poll_play_proto_gen/public.dart';
 class GamesClient {
   late GamesServiceClient _client;
 
-  GamesClient(List<String> address) {
-    final channel = GrpcOrGrpcWebClientChannel.toSingleEndpoint(
-        host: address[0], port: int.parse(address[1]), transportSecure: GlobalConfig().secureTransport);
+  GamesClient(ClientChannel channel) {
     _client = GamesServiceClient(channel);
   }