+++ lib/grpc/user.dart
@@ -1,5 +1,4 @@
-import 'package:grpc/grpc_or_grpcweb.dart';
-import 'package:poll_and_play/config.dart';
+import 'package:grpc/grpc.dart';
 import 'package:poll_and_play/grpc/authenticator.dart';
 import 'package:poll_play_proto_gen/google.dart';
 import 'package:poll_play_proto_gen/public.dart';
@@ -7,9 +6,7 @@ import 'package:poll_play_proto_gen/public.dart';
 class UserClient {
   late UserServiceClient _client;
 
-  UserClient(List<String> address) {
-    final channel = GrpcOrGrpcWebClientChannel.toSingleEndpoint(
-        host: address[0], port: int.parse(address[1]), transportSecure: GlobalConfig().secureTransport);
+  UserClient(ClientChannel channel) {
     _client = UserServiceClient(channel);
   }