+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Added:
 - foreground notifications support
 - user device registration
+- move to single grpc channel
 
 ## 0.4.0 (2024-03-23)